Fixed warnings

+++ Launchy_QT/plugins/runner/runner.cpp
@@ -99,7 +99,7 @@ void runnerPlugin::getCatalog(QList<CatItem>* items)
 
 void runnerPlugin::getResults(QList<InputData>* id, QList<CatItem>* results)
 {
-	if (id->count() > 1 && id->first().getTopResult().id == HASH_runner) {
+    if ( id->count() > 1 && (unsigned int)id->first().getTopResult().id == HASH_runner) {
 		const QString & text = id->last().getText();
 		// This is user search text, create an entry for it
 		results->push_front(CatItem(text, text, HASH_runner, getIcon()));
